fix test failure on CI pipeline
